vector_service/requirements.txt
fastapi
uvicorn
qdrant-client>=1.7.0
sqlalchemy
httpx
pydantic-settings
pandas
python-multipart
